+++ test/support/shrink_quality_support.ex
@@ -232,16 +232,16 @@ defmodule PropertyDamage.Test.ShrinkQuality.CorrectAdapter do
   end
 
   @impl true
-  def execute(%PutKey{key: key, value: value}, %{store: pid}) do
+  def execute(%PutKey{key: key, value: value}, %{store: pid}, _runtime) do
     Store.put(pid, key, value)
     {:ok, [%EntryPut{key: key, value: value}]}
   end
 
-  def execute(%GetKey{key: key}, %{store: pid}) do
+  def execute(%GetKey{key: key}, %{store: pid}, _runtime) do
     {:ok, [%EntryRead{key: key, value: Store.get(pid, key)}]}
   end
 
-  def execute(%DelKey{key: key}, %{store: pid}) do
+  def execute(%DelKey{key: key}, %{store: pid}, _runtime) do
     Store.delete(pid, key)
     {:ok, [%EntryDeleted{key: key}]}
   end